move machine account key generation func
