+++ cmd/bootstrap/cmd/key.go
@@ -2,13 +2,13 @@ package cmd
 
 import (
 	"fmt"
+	"github.com/onflow/flow-go/cmd/bootstrap/utils"
 	"net"
 	"strconv"
 
 	"github.com/multiformats/go-multiaddr"
 	"github.com/spf13/cobra"
 
-	"github.com/onflow/flow-go/cmd/bootstrap/run"
 	"github.com/onflow/flow-go/crypto"
 	model "github.com/onflow/flow-go/model/bootstrap"
 	"github.com/onflow/flow-go/model/flow"
@@ -88,7 +88,7 @@ func generateKeys() (crypto.PrivateKey, crypto.PrivateKey, crypto.PrivateKey, er
 
 	log.Debug().Msg("will generate networking key")
 	networkSeed := validateSeed(flagNetworkSeed)
-	networkKey, err := run.GenerateNetworkingKey(networkSeed)
+	networkKey, err := utils.GenerateNetworkingKey(networkSeed)
 	if err != nil {
 		return nil, nil, nil, fmt.Errorf("could not generate networking key: %w", err)
 	}
@@ -96,7 +96,7 @@ func generateKeys() (crypto.PrivateKey, crypto.PrivateKey, crypto.PrivateKey, er
 
 	log.Debug().Msg("will generate staking key")
 	stakingSeed := validateSeed(flagStakingSeed)
-	stakingKey, err := run.GenerateStakingKey(stakingSeed)
+	stakingKey, err := utils.GenerateStakingKey(stakingSeed)
 	if err != nil {
 		return nil, nil, nil, fmt.Errorf("could not generate staking key: %w", err)
 	}
@@ -104,7 +104,7 @@ func generateKeys() (crypto.PrivateKey, crypto.PrivateKey, crypto.PrivateKey, er
 
 	log.Debug().Msg("will generate machine account key")
 	machineSeed := validateSeed(flagMachineSeed)
-	machineKey, err := run.GenerateMachineAccountKey(machineSeed)
+	machineKey, err := utils.GenerateMachineAccountKey(machineSeed)
 	if err != nil {
 		return nil, nil, nil, fmt.Errorf("could not generate machine key: %w", err)
 	}
