+++ cmd/bootstrap/cmd/keys.go
@@ -2,9 +2,9 @@ package cmd
 
 import (
 	sdkcrypto "github.com/onflow/flow-go-sdk/crypto"
+	"github.com/onflow/flow-go/cmd/bootstrap/utils"
 	"github.com/onflow/flow-go/model/flow/order"
 
-	"github.com/onflow/flow-go/cmd/bootstrap/run"
 	"github.com/onflow/flow-go/crypto"
 	model "github.com/onflow/flow-go/model/bootstrap"
 	"github.com/onflow/flow-go/model/encodable"
@@ -25,14 +25,14 @@ func genNetworkAndStakingKeys() []model.NodeInfo {
 	log.Debug().Msg("all node addresses are unique")
 
 	log.Debug().Msgf("will generate %v networking keys for nodes in config", nodes)
-	networkKeys, err := run.GenerateNetworkingKeys(nodes, generateRandomSeeds(nodes))
+	networkKeys, err := utils.GenerateNetworkingKeys(nodes, generateRandomSeeds(nodes))
 	if err != nil {
 		log.Fatal().Err(err).Msg("cannot generate networking keys")
 	}
 	log.Info().Msgf("generated %v networking keys for nodes in config", nodes)
 
 	log.Debug().Msgf("will generate %v staking keys for nodes in config", nodes)
-	stakingKeys, err := run.GenerateStakingKeys(nodes, generateRandomSeeds(nodes))
+	stakingKeys, err := utils.GenerateStakingKeys(nodes, generateRandomSeeds(nodes))
 	if err != nil {
 		log.Fatal().Err(err).Msg("cannot generate staking keys")
 	}
