+++ cmd/bootstrap/utils/key_generation.go
@@ -1,4 +1,4 @@
-package run
+package utils
 
 import (
 	"fmt"
