+++ cmd/bootstrap/utils/key_generation_test.go
@@ -1,4 +1,4 @@
-package run
+package utils
 
 import (
 	"fmt"
